+++ phase2/src/ATM/Accounts/DebtAccount.java
@@ -58,7 +58,7 @@ public abstract class DebtAccount extends BasicAccount {
      * @param amount The given amount of money
      */
     @Override
-    public void deposit(int amount){
+    public void deposit(double amount){
         this.balance -= amount;
     }
