+++ phase2/src/ATM/Accounts/TransferTypes/Depositable.java
@@ -3,5 +3,5 @@ package ATM.Accounts.TransferTypes;
  * The interface for depositable of the accounts
  */
 public interface Depositable {
-    void deposit(int amount);
+    void deposit(double amount);
 }
