+++ phase2/src/ATM/Accounts/TransferTypes/Withdrawable.java
@@ -2,5 +2,5 @@ package ATM.Accounts.TransferTypes;
 
 /** A withdrawable account can be withdraw amount of money*/
 public interface Withdrawable {
-      void withdraw(int amount);
+      void withdraw(double amount);
 }
